requirements.txt
tavily-python>=0.1.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi-mcp>=0.1.0
pytest>=7.0.0
httpx>=0.24.0
tenacity>=8.2.0
html2text>=2020.1.16 